+++ unicode-data-names/Changelog.md
@@ -5,6 +5,8 @@
 - Improve performance.
 - Added opional support for `ByteString` API.
   Use the package flag `has-bytestring` to enable it.
+- Added opional support for `Text` API.
+  Use the package flag `has-text` to enable it.
 
 ## 0.2.0 (September 2022)